+++ frontend/csyllabus/src/app/core/explorer/explorer/explorer.component.css
