+++ frontend/csyllabus/src/app/core/explorer/explorer/explorer.module.ts
@@ -1,7 +1,7 @@
 import { ExplorerComponent } from './explorer.component';
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { AngularMaterialModule } from './../../angular-material/angular-material.module';
+import { AngularMaterialModule } from './../../../angular-material/angular-material.module';
 
 @NgModule({
